feature: #1917 Restructure app detail folders
