+++ packages/developer-portal/src/components/pages/app-detail/__tests__/app-content.test.tsx
