+++ packages/developer-portal/src/components/pages/app-detail/__tests__/app-header.test.tsx
@@ -1,7 +1,7 @@
 import * as React from 'react'
 import { shallow } from 'enzyme'
 import { appDetailDataStub } from '@/sagas/__stubs__/app-detail'
-import AppHeader, { AppHeaderProps } from '../ui-app-header'
+import AppHeader, { AppHeaderProps } from '../app-header'
 
 const mockProps: AppHeaderProps = {
   appDetailData: {
