+++ packages/developer-portal/src/components/pages/app-detail/__tests__/manage-app.test.tsx
