+++ packages/developer-portal/src/components/pages/app-detail/app-detail.tsx
@@ -10,12 +10,12 @@ import { History } from 'history'
 import { selectAppDetailState, selectAppDetailData, selectAppDetailLoading } from '@/selector/developer-app-detail'
 import { selectInstallAppLoading } from '@/selector/installations'
 import { Loader, Grid, GridItem, Section } from '@reapit/elements'
-import AppHeader from '@/components/pages/app-detail/common/ui-app-header'
+import AppHeader from '@/components/pages/app-detail/app-header'
 import styles from '@/styles/blocks/standalone-app-detail.scss?mod'
 import routes from '@/constants/routes'
 import { getDesktopIntegrationTypes } from '@/utils/get-desktop-integration-types'
 import useReactResponsive from '@/components/hooks/use-react-responsive'
-import { BackToAppsSection } from '../common/ui-sections'
+import { BackToAppsSection } from './app-sections'
 import AppContent from './app-content'
 
 export type DeveloperAppDetailProps = {}
