+++ packages/developer-portal/src/components/pages/app-detail/helpers.tsx
