+++ packages/developer-portal/src/components/pages/app-detail/index.ts
